requirements.txt
feedparser
psycopg2-binary
google-generativeai
requests
python-dotenv
